LU-14865 utils: llog_reader.c printf type mismatch
Add (unsigned long long) cast to results of
__le64_to_cpu so that it matches the formatting (%llu)
of the enclosing printf call.

Build log message:
"llog_reader.c:887:9: error: format '%llu' expects
argument of type 'long long unsigned int', but
argument 3 has type '__u64' [-Werror=format=]"

Test-Parameters: trivial
Fixes: 9962d6f84db5 LU-14617 utils: llog_reader updatelog support
Reviewed-on: https://review.whamcloud.com/44346
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/utils/llog_reader.c
@@ -886,8 +886,8 @@ void print_update_rec(struct llog_update_record *lur)
 	up_count = __le32_to_cpu(rec->ur_update_count);
 	pm_count = __le32_to_cpu(rec->ur_param_count);
 	printf("updatelog record master_transno:%llu batchid:%llu flags:0x%x u_index:%d u_count:%d p_count:%d\n",
-	       __le64_to_cpu(rec->ur_master_transno),
-	       __le64_to_cpu(rec->ur_batchid),
+	       (unsigned long long)__le64_to_cpu(rec->ur_master_transno),
+	       (unsigned long long)__le64_to_cpu(rec->ur_batchid),
 	       __le32_to_cpu(rec->ur_flags),
 	       __le32_to_cpu(rec->ur_index),
 	       up_count,
